Bug fixes in model, fixing V2 backwards compatibility
git-svn-id: svn+ssh://mammoth.bio.cam.ac.uk/data/ccpn/svnhome/ccpnmodel/trunk@223 335f5335-426b-4637-abd2-6528e8beab9b

+++ testing/Test_compatibility.py
@@ -178,8 +178,11 @@ def testProject(target, outDir):
     if not newPath.endswith(corePath.CCPN_DIRECTORY_SUFFIX):
       newPath += corePath.CCPN_DIRECTORY_SUFFIX
     logger.info('### saving %s to %s' % (ccpnProject.name, newPath))
-    apiIo.saveProject(ccpnProject, newPath=newPath, newProjectName=ccpnProject.name,
-                       overwriteExisting=True)
+    # apiIo.saveProject(ccpnProject, newPath=newPath, newProjectName=ccpnProject.name,
+    #                    overwriteExisting=True)
+    result = apiIo.saveProject(ccpnProject, newPath=newPath, overwriteExisting=True)
+    if not result:
+      raise ValueError("CCPN save failed - apiIo.saveProject did not return Trus")
     shutil.make_archive(newPath, 'gztar', newPath)
     shutil.rmtree(newPath)
     t4 = time.time()
